#190 Refactored ship system with common template architecture
