+++ apps/publish-frt/base/src/builders/templates/mmoomm/playcanvas/handlers/shared/shipSystems/index.ts
@@ -1,0 +1,6 @@
+// Universo Platformo | MMOOMM Ship Systems
+// Central export for all ship-related shared templates
+
+export * from './shipControllerTemplate'
+export * from './cameraControllerTemplate'
+export * from './laserMiningTemplate'
